Day 17

+++ Cpp/solutions/day17.hpp
@@ -5,10 +5,12 @@
 #include <fstream>
 #include <string>
 #include <vector>
+#include <utility>
+#include <map>
 
 namespace day17 {
     int part1(std::string input);
-    int part2(std::string input);
+    long long part2(std::string input);
 }
 
 #endif // DAY17_HPP
